Moved EventHandler and BatchRequest into more appropriate locations. Made SdkVersionMiddleware configurable.
